% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NNCTFunctions.r
\name{funsZcell.spec}
\alias{funsZcell.spec}
\alias{cell.spec.ct}
\alias{cell.spec}
\title{Cell-specific Z Tests of Segregation for NNCTs}
\usage{
cell.spec.ct(
  ct,
  covN,
  type,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)

cell.spec(
  dat,
  lab,
  type,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{ct}{A nearest neighbor contingency table, used in \code{Zcell.spec.ct} only}

\item{covN}{The \eqn{k^2 \times k^2} covariance matrix of row-wise vectorized entries of NNCT, \code{ct} ;
used in \code{Zcell.spec.ct} only.}

\item{type}{The type of the cell-specific test with no default.
Takes on values \code{"dixon"} or \code{"nnct"} for Dixon's cell-specific tests and \code{"I"}-\code{"IV"} for types I-IV cell-specific
tests (or equivalently \code{1-6}, respectively).}

\item{alternative}{Type of the alternative hypothesis in the test, one of \code{"two.sided"}, \code{"less"} or \code{"greater"}.}

\item{conf.level}{Level of the upper and lower confidence limits, default is \code{0.95}, for the \eqn{N_{ij}} or \eqn{T_{ij}} values}

\item{dat}{The data set in one or higher dimensions, each row corresponds to a data point,
used in \code{Zcell.spec} only}

\item{lab}{The \code{vector} of class labels (numerical or categorical), used in \code{Zcell.spec} only}

\item{\dots}{are for further arguments, such as \code{method} and \code{p}, passed to the \code{\link[stats]{dist}} function, used in \code{Zcell.spec} only}
}
\value{
A \code{list} with the elements
\item{statistic}{The \code{matrix} of cell-specific test statistics}
\item{stat.names}{Name of the test statistics}
\item{p.value}{The \code{matrix} of \eqn{p}-values for the hypothesis test for the corresponding alternative}
\item{LCL,UCL}{Matrix of Lower and Upper Confidence Levels for the \eqn{N_{ij}} or \eqn{T_{ij}} values at the given confidence
level \code{conf.level} and depends on the type of \code{alternative}.}
\item{conf.int}{The confidence interval for the estimates, it is \code{NULL} here, since we provide the \code{UCL} and \code{LCL}
in \code{matrix} form.}
\item{cnf.lvl}{Level of the upper and lower confidence limits of the entries, provided in \code{conf.level}.}
\item{estimate}{Estimates of the parameters, NNCT or TCT, i.e., matrix of the observed \eqn{N_{ij}} or \eqn{T_{ij}} values
which is NNCT or TCT, respectively.}
\item{est.name,est.name2}{Names of the estimates, both are same in this function}
\item{null.value}{Matrix of hypothesized null values for the parameters which are expected values of the 
the null \eqn{N_{ij}} values in an NNCT or \eqn{T_{ij}} values in an TCT.}
\item{null.name}{Name of the null values}
\item{alternative}{Type of the alternative hypothesis in the test, one of \code{"two.sided"}, \code{"less"} or \code{"greater"}}
\item{method}{Description of the hypothesis test}
\item{ct.name}{Name of the contingency table, \code{ct}, returned by \code{Zcell.spec.ct} only}
\item{data.name}{Name of the data set, \code{dat}, returned by \code{Zcell.spec} only}
}
\description{
Two functions: \code{Zcell.spec.ct} and \code{Zcell.spec}.

All functions are objects of class \code{"cellhtest"} but with different arguments (see the parameter list below).
Each one performs hypothesis tests of deviations of 
entries of NNCT or types I-IV TCTs from the expected values under RL or CSR for each entry.
The test for each entry \eqn{i,j} is based on the normal approximation of the corresponding \eqn{T_{ij}} value
and are due to \insertCite{dixon:NNCTEco2002;textual}{nnspat}
and \insertCite{ceyhan:jkss-posthoc-2017;textual}{nnspat}, respectively.

The \code{type="dixon"} or \code{"nnct"} refers to Dixon's cell-specific test of segregation, and
\code{type="I"}-\code{"IV"} refers to types I-IV cell-specific tests, respectively.

Each function yields a contingency table of the test statistics, \eqn{p}-values for the corresponding 
alternative, expected values (i.e. null value(s)), lower and upper confidence levels and sample estimates (i.e. observed values)
for the \eqn{N_{ij}} or \eqn{T_{ij}} values and also names of the test statistics, estimates, null values and the method and
the data set used.

The null hypothesis for each entry \eqn{i,j} is that the corresponding value \eqn{N_{ij}} or \eqn{T_{ij}} is equal to the 
expected value under RL or CSR.

See also
(\insertCite{dixon:1994,dixon:NNCTEco2002,ceyhan:eest-2010,ceyhan:jkss-posthoc-2017;textual}{nnspat})
and the references therein.
}
\examples{
n<-20  #or try sample(1:20,1)
Y<-matrix(runif(3*n),ncol=3)
ipd<-ipd.mat(Y)
cls<-sample(1:2,n,replace = TRUE)  #or try cls<-rep(1:2,c(10,10))
ct<-nnct(ipd,cls)

W<-Wmat(ipd)
Qv<-Qvec(W)$q
Rv<-Rval(W)
varN<-var.nnct(ct,Qv,Rv)
covN<-cov.nnct(ct,varN,Qv,Rv)

type<-"IV" #"dixon" #try also "nnct", "I", "II", "III", and "IV"
cell.spec(Y,cls,type)
cell.spec(Y,cls,type,alt="g")

cell.spec.ct(ct,covN,type)
cell.spec.ct(ct,covN,type="II",alt="g")

cell.spec(Y,cls,type,method="max")

#cls as a factor
na<-floor(n/2); nb<-n-na
fcls<-rep(c("a","b"),c(na,nb))
cell.spec(Y,cls,type="I")

#############
n<-40
Y<-matrix(runif(3*n),ncol=3)
ipd<-ipd.mat(Y)
cls<-sample(1:4,n,replace = TRUE)  #or try cls<-rep(1:2,c(10,10))
ct<-nnct(ipd,cls)

W<-Wmat(ipd)
Qv<-Qvec(W)$q
Rv<-Rval(W)
varN<-var.nnct(ct,Qv,Rv)
covN<-cov.nnct(ct,varN,Qv,Rv)

cell.spec(Y,cls,type)
cell.spec.ct(ct,covN,type)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link{Zcell.nnct.ct}}, \code{\link{Zcell.nnct}}, \code{\link{Zcell.tct.ct}} 
and \code{\link{Zcell.tct}}
}
\author{
Elvan Ceyhan
}
