% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4NNCTClasses.R
\name{plot.Clusters}
\alias{plot.Clusters}
\title{Plot a \code{Clusters} object}
\usage{
\method{plot}{Clusters}(x, asp = NA, xlab = "x", ylab = "y", ...)
}
\arguments{
\item{x}{Object of class \code{Clusters}.}

\item{asp}{A numeric value, giving the aspect ratio for y axis to x-axis y/x (default is \code{NA}),
see the official help for \code{asp} by typing "? asp".}

\item{xlab, ylab}{Titles for the x and y axes, respectively (default is \code{xlab}="x" and \code{ylab}="y").}

\item{\dots}{Additional parameters for \code{plot}.}
}
\value{
None
}
\description{
Plots the points generated from the pattern (color coded for each class) together with the
study window
}
\examples{
#TBF

}
