% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4NNCTClasses.r
\name{print.Chisqtest}
\alias{print.Chisqtest}
\title{Print a summary of a \code{Chisqtest} object}
\usage{
\method{print}{Chisqtest}(x, digits = getOption("digits"), prefix = "\\t", ...)
}
\arguments{
\item{x}{object of class "\code{summary.Chisqtest}"}

\item{digits}{number of significant digits to be used.}

\item{prefix}{string, passed to \code{\link[base]{strwrap}} for displaying the method component 
of the \code{classhtest} object.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Printing objects of class "\code{Chisqtest}" by simple \code{\link[base]{print}} methods.
}
