% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4NNCTClasses.r
\name{print.Clusters}
\alias{print.Clusters}
\title{Print a \code{Clusters} object}
\usage{
\method{print}{Clusters}(x, ...)
}
\arguments{
\item{x}{A \code{Clusters} object.}

\item{\dots}{Additional arguments for the S3 method '\code{print}'.}
}
\value{
The \code{call} of the object of class '\code{Clusters}'
and also the \code{type} (or description) of the pattern).
}
\description{
Prints the \code{call} of the object of class '\code{Clusters}'
and also the \code{type} (or description) of the pattern).
}
\examples{
#TBF (to be filled)

}
\seealso{
\code{\link{summary.Clusters}}, \code{\link{print.summary.Clusters}}, and \code{\link{plot.Clusters}}
}
