% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4NNCTClasses.R
\name{print.SpatPatterns}
\alias{print.SpatPatterns}
\title{Print a \code{SpatPatterns} object}
\usage{
\method{print}{SpatPatterns}(x, ...)
}
\arguments{
\item{x}{A \code{SpatPatterns} object.}

\item{\dots}{Additional arguments for the S3 method '\code{print}'.}
}
\value{
The \code{call} of the object of class '\code{SpatPatterns}'
and also the \code{type} (or description) of the pattern).
}
\description{
Prints the \code{call} of the object of class '\code{SpatPatterns}'
and also the \code{type} (or description) of the pattern).
}
\examples{
#TBF (to be filled)

}
\seealso{
\code{\link{summary.SpatPatterns}}, \code{\link{print.summary.SpatPatterns}}, and \code{\link{plot.SpatPatterns}}
}
