% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4NNCTClasses.R
\name{print.summary.Clusters}
\alias{print.summary.Clusters}
\title{Print a summary of a \code{Clusters} object}
\usage{
\method{print}{summary.Clusters}(x, ...)
}
\arguments{
\item{x}{object of class "\code{summary.Clusters}", generated by \code{summary.Clusters}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.Clusters}}, \code{\link{summary.Clusters}}, and \code{\link{plot.Clusters}}
}
