% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4NNCTClasses.R
\name{print.summary.SpatPatterns}
\alias{print.summary.SpatPatterns}
\title{Print a summary of a \code{SpatPatterns} object}
\usage{
\method{print}{summary.SpatPatterns}(x, ...)
}
\arguments{
\item{x}{object of class "\code{summary.SpatPatterns}", generated by \code{summary.SpatPatterns}.}

\item{\dots}{Additional parameters for \code{print}.}
}
\value{
None
}
\description{
Prints some information about the \code{object}.
}
\seealso{
\code{\link{print.SpatPatterns}}, \code{\link{summary.SpatPatterns}}, and \code{\link{plot.SpatPatterns}}
}
