% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxRFuncs4NNCTClasses.R
\name{summary.Clusters}
\alias{summary.Clusters}
\title{Return a summary of a \code{Clusters} object}
\usage{
\method{summary}{Clusters}(object, ...)
}
\arguments{
\item{object}{Object of class \code{Clusters}.}

\item{\dots}{Additional parameters for \code{summary}.}
}
\value{
The \code{call} of the object of class '\code{Clusters}', the \code{type} of the pattern, \code{parameters} of the pattern,
study window, some sample points from the generated pattern, reference points (if any for the bivariate pattern),
and number of points for each class
}
\description{
Returns the below information about the \code{object}:

\code{call} of the function defining the \code{object}, the \code{type} of the pattern, \code{parameters} of the pattern,
study window, some sample points from the generated pattern, reference points (if any for the bivariate pattern),
and number of points for each class
}
\examples{
#TBF
}
